FLEETSW-5461: fix lidar profiles
Two profiles, the FUSA profile and the LB profile didn't have enough bits in the RANGE fields to accommodate the left shift. Changing the fields to be 32-bit from 16-bit fixes this.

FLEETSW-5461: apply proper scaling to FUSA NEAR_IR field

FLEETSW-5461: fix LB profile NEAR_IR field invalid read

FLEETSW-5461: fix UT associated with FUSA profile

Fix snapshot test, fix bug with default FUSA fields in LidarScan

Clang format

FLEETSW-5461: update expected data digest for LB profile after profile fixes

FLEETSW-5461: update CHANGELOG to indicate breaking changes (changes in LB field widths)

+++ ouster_client/src/lidar_scan.cpp
@@ -90,7 +90,7 @@ static const Table<ChanField, ChanFieldType, 5> five_word_slots{{
 static const Table<ChanField, ChanFieldType, 7> fusa_two_word_slots{{
     {ChanField::RANGE, ChanFieldType::UINT32},
     {ChanField::REFLECTIVITY, ChanFieldType::UINT8},
-    {ChanField::NEAR_IR, ChanFieldType::UINT8},
+    {ChanField::NEAR_IR, ChanFieldType::UINT16},
     {ChanField::RANGE2, ChanFieldType::UINT32},
     {ChanField::REFLECTIVITY2, ChanFieldType::UINT8},
     {ChanField::RAW32_WORD1, ChanFieldType::UINT32},
